fix(ui): remove self-reexport for useTokenEditor
